AddPoints parseFloat instead of int

+++ commands/addPoints.js
@@ -36,7 +36,7 @@ function addAmountOfPointsToSb(message, points, users, member, ranking_titles, r
 
 module.exports = (guildId, scores, message,words, rankings_links, rankings_titles) => {
     if (message.member.roles.cache.find(r => r.name === "Admin" || r.name === "b! point giver")) {
-        let points = parseInt(words[2])
+        let points = parseFloat(words[2])
 
         if (isNaN(points))
             return message.reply("Error! Mention how much points you want to add to mentioned people.\nTry : b!add <ranking> <number of points> <@nobody> <@some cool guy> ... <@another one>")
